refactor: add and using a wrapper component for Bar chart metrics
