+++ src/components/MetricCharts/index.ts
@@ -1,3 +1,4 @@
 import { LineChartMetric } from '@/components/MetricCharts/LineChartMetric';
+import StackedBarCharMetric from '@/components/MetricCharts/StackedBarCharMetric';
 
-export { LineChartMetric };
+export { LineChartMetric, StackedBarCharMetric };
